+++ example/lib/src/view/home/page_01.dart
@@ -205,8 +205,7 @@ class Page1State extends StateX<Page1> with EventsStateMixin<Page1> {
     if (stack != null && stack.toString().contains('handleTap')) {
       // Increment the count like no error occurred
       count++;
-
-      /// Look how this Controller has access to this State object!
+      // Look how this Controller has access to this State object!
       controller?.setState(() {});
     }